SL-81 Response body validation
