+++ packages/http/src/validator/index.ts
@@ -1,10 +1,15 @@
-import { header as headerDeserializerRegistry } from './deserializer';
+import {
+  header as headerDeserializerRegistry,
+  query as queryDeserializerRegistry,
+} from './deserializer';
+import { HttpBodyValidator } from './helpers/HttpBodyValidator';
 import { HttpHeadersValidator } from './helpers/HttpHeadersValidator';
-import { HttpRequestBodyValidator } from './helpers/HttpRequestBodyValidator';
+import { HttpQueryValidator } from './helpers/HttpQueryValidator';
 import { HttpValidator } from './HttpValidator';
 import { validatorRegistry } from './registry';
 
 export const validator = new HttpValidator(
-  new HttpRequestBodyValidator(validatorRegistry),
-  new HttpHeadersValidator(headerDeserializerRegistry)
+  new HttpBodyValidator(validatorRegistry),
+  new HttpHeadersValidator(headerDeserializerRegistry),
+  new HttpQueryValidator(queryDeserializerRegistry)
 );
